refactor-project (#49)
* ♻️ Refactor project as a multi-modules build

* ♻️ Module "kotlin-testing" now contains all the tests

* ♻️ Module "kotlin-codegen" now contains all the code-generation stuff

* 🛠 Fix GitHub Action workflow

* 🛠 Remove often failing test

 # Kotlin-faker - Port of a popular ruby faker gem written in Kotlin.
Name: Earlene Adams Torp Fahey
Address: 325 Alva Ford Suite 216 MT Guyana
Phone number: 940.053.2366 x577

> Task :kotlin-jvm:run FAILED
Error: Exception in thread "main" java.util.NoSuchElementException: Secondary key 'el salvador' not found.

https://github.com/LouisCAD/kotlin-libraries-playground/pull/49/checks?check_run_id=1224763768

* ✚ EditorConfig file

* 🐛 kotlinOptions.freeCompilerArgs

* Fast plugins can stay in "kotlin-jvm"

* PR feedback

* PR feedback
