+++ gradle.properties
@@ -1,3 +1,4 @@
 # https://dev.to/jmfayard/configuring-gradle-with-gradle-properties-211k
 kotlin.code.style=official
 org.gradle.caching=true
+org.gradle.parallel=true
