+++ kotlin-codegen/src/main/kotlin/playground/utils.kt
@@ -1,0 +1,6 @@
+package playground
+
+infix fun <T : Any?> T.shouldBe(expected: T) {
+    println("Test: $expected")
+    check(this == expected) { "Test failed!\nWanted: $expected\nGot:    $this" }
+}
