+++ kotlin-jvm/src/main/kotlin/playground/Apollo.kt
@@ -7,42 +7,43 @@ import com.apollographql.apollo.coroutines.await
 import com.apollographql.apollo.exception.ApolloException
 import kotlinx.coroutines.runBlocking
 import launches.GetLaunchesQuery
+import okhttp3.Dispatcher
 import okhttp3.OkHttpClient
 import java.util.concurrent.Executors
-import java.util.concurrent.ThreadPoolExecutor
 
 
 fun main() {
-  println()
-  println("# apollographql/Apollo - Typesafe GraphQL in Kotlin")
+    println()
+    println("# apollographql/Apollo - Typesafe GraphQL in Kotlin")
 
-  val okHttpClient = OkHttpClient()
-  val dispatcher = Executors.newCachedThreadPool()
+    val executorService = Executors.newCachedThreadPool()
+    val dispatcher = Dispatcher(executorService)
+    val okHttpClient = OkHttpClient.Builder().dispatcher(dispatcher).build()
 
-  val apolloClient = ApolloClient.builder()
-      .serverUrl("https://apollo-fullstack-tutorial.herokuapp.com/graphql")
-      .okHttpClient(okHttpClient)
-      .dispatcher(dispatcher)
-      .build()
+    val apolloClient = ApolloClient.builder()
+        .serverUrl("https://apollo-fullstack-tutorial.herokuapp.com/graphql")
+        .okHttpClient(okHttpClient)
+        .dispatcher(executorService)
+        .build()
 
-  runBlocking {
-    val response = try {
-       apolloClient.query(GetLaunchesQuery()).await()
-    } catch (e: ApolloException) {
-      println("network error:")
-      e.printStackTrace()
-      return@runBlocking
+    runBlocking {
+        val response = try {
+            apolloClient.query(GetLaunchesQuery()).await()
+        } catch (e: ApolloException) {
+            println("network error:")
+            e.printStackTrace()
+            return@runBlocking
+        }
+
+        response.data?.launches?.launches?.forEach {
+            println("Mission ${it?.mission?.name} launched from site ${it?.site} ")
+        }
     }
 
-    response.data?.launches?.launches?.forEach {
-      println("Mission ${it?.mission?.name} launched from site ${it?.site} ")
-    }
-  }
-
-  /**
-   * Shutdown any running threadpool to make the command terminate
-   * (The JVM waits for all threads to terminate before exiting by default)
-   */
-  dispatcher.shutdown()
-  okHttpClient.dispatcher.executorService.shutdown()
+    /**
+     * Shutdown any running threadpool to make the command terminate
+     * (The JVM waits for all threads to terminate before exiting by default)
+     */
+    executorService.shutdown()
+    dispatcher.executorService.shutdown()
 }
