+++ kotlin-jvm/src/main/kotlin/playground/KtorClient.kt
@@ -6,18 +6,12 @@ import io.ktor.client.*
 import io.ktor.client.engine.okhttp.*
 import io.ktor.client.features.json.*
 import io.ktor.client.request.*
-import io.ktor.http.*
-import io.ktor.util.*
 import kotlinx.coroutines.runBlocking
-import kotlinx.serialization.ExperimentalSerializationApi
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.Json
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
+import playground.kotlinx.serialization.HttpBinGet
 import playground.ktor.client.Network.API_URL
 import playground.shouldBe
-import retrofit2.*
-import retrofit2.http.QueryMap
 
 /**
  * Ktor Http Clients
@@ -43,7 +37,7 @@ class KtorHttpbinApi(
 ) {
 
     suspend fun get(vararg params: Pair<String, String>) =
-            client.get<HttpbinGet>("$API_URL/get") {
+            client.get<HttpBinGet>("$API_URL/get") {
                 params.forEach {
                     url.parameters.append(it.first, it.second)
                 }
@@ -69,10 +63,3 @@ object Network {
     }
 }
 
-@Serializable
-data class HttpbinGet(
-        val args: Map<String, String> = emptyMap(),
-        val headers: Map<String, String> = emptyMap(),
-        val origin: String,
-        val url: String
-)
