+++ kotlin-jvm/src/main/kotlin/playground/_main.kt
@@ -32,7 +32,4 @@ fun main() {
      */
 }
 
-infix fun <T : Any?> T.shouldBe(expected: T) {
-    println("Test: $expected")
-    check(this == expected) { "Test failed!\nWanted: $expected\nGot:    $this" }
-}
+
