+++ kotlin-testing/src/test/kotlin/framework/common/UseCase.kt
