+++ kotlin-testing/src/test/kotlin/framework/junit5/UseCaseTest.kt
