+++ kotlin-testing/src/test/kotlin/framework/kotest/DistanceConverterParameterizedTest.kt
