+++ kotlin-testing/src/test/kotlin/framework/spek/UseCaseTest.kt
