+++ kotlin-testing/src/test/kotlin/framework/strikt/DistanceConverterParameterizedTest.kt
