+++ kotlin-testing/src/test/kotlin/framework/strikt/UseCaseTest.kt
