+++ settings.gradle.kts
@@ -1,10 +1,10 @@
 import de.fayard.refreshVersions.bootstrapRefreshVersions
 
 buildscript {
-    repositories { mavenLocal() ; gradlePluginPortal() }
+    repositories { mavenLocal(); gradlePluginPortal() }
     val useSnapshot = false
     dependencies.classpath(
-            if (useSnapshot) "de.fayard.refreshVersions:refreshVersions:0.9.6-SNAPSHOT" else "de.fayard.refreshVersions:refreshVersions:0.9.5"
+        if (useSnapshot) "de.fayard.refreshVersions:refreshVersions:0.9.6-SNAPSHOT" else "de.fayard.refreshVersions:refreshVersions:0.9.5"
     )
 }
 
@@ -24,3 +24,7 @@ gradleEnterprise {
 rootProject.name = "kotlin-libraries-playground"
 
 bootstrapRefreshVersions()
+
+include("kotlin-jvm")
+include("kotlin-testing")
+include("kotlin-codegen")
